Force results files to follow umask, even when aborting a job.
Many log files are not originally world-readable, and rsync -a means
they will be unreadable in the web ui, with e.g. autotest running as
user "autotest" and a web server running as "www-data".

Also preserve symlinks in the aborted job case, as that was done for a
non-aborting job already.